+++ src/metatrain/experimental/flashmd_symplectic/documentation.py
@@ -9,9 +9,9 @@ The symplectic variant of :ref:`FlashMD <arch-flashmd>`.
 {{SECTION_DEFAULT_HYPERS}}
 """
 
-from typing import Literal, Optional, Union
+from typing import Literal, Optional
 
-from typing_extensions import TypedDict
+from typing_extensions import NotRequired, TypedDict
 
 from metatrain.pet.modules.finetuning import FinetuneHypers, NoFinetuneHypers
 from metatrain.utils.hypers import init_with_defaults
@@ -117,10 +117,10 @@ class TrainerHypers(TypedDict):
     module will be used. These correspond to masses averaged over the natural
     isotopic abundance of each element.
     """
-    distributed: Union[bool, Literal["auto"]] = "auto"
-    """Whether to use distributed training. The default ``"auto"`` enables
-    distributed training when running under more than one SLURM task.
-    Explicit booleans override the detection and are deprecated."""
+    distributed: NotRequired[bool]
+    """Whether to use distributed training. When not set, distributed training
+    is enabled automatically when running under more than one SLURM task.
+    Setting this option explicitly is deprecated."""
     distributed_port: int = 39591
     """Port for DDP communication"""
     batch_size: int = 16